BUG: A ticket must not be assigned to anyone be 'in progress'.
The lookup fails when a ticket is put into 'in progress' but does not
have a ticket assignee set. A valid redmine case but not very
practical in the real world. If this combination is found the assignee
name will now be shown as '??' to hilight the problem in channel.